requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.3.0
supabase==2.5.0
onnxruntime==1.24.1
Pillow==11.1.0
numpy==1.26.4
python-magic==0.4.27
python-jose[cryptography]==3.3.0
slowapi==0.1.9
huggingface_hub==0.23.0
sentry-sdk[fastapi]==2.4.0
